+++ packages/template/src/lib/stack-app.ts
@@ -1036,10 +1036,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       },
       // END_PLATFORM react-like
       async createTeam(data: TeamCreateOptions) {
-        const crud = await app._interface.createClientTeam({
-          ...teamCreateOptionsToCrud(data),
-          creator_user_id: 'me',
-        }, session);
+        const crud = await app._interface.createClientTeam(teamCreateOptionsToCrud(data, 'me'), session);
         await app._currentUserTeamsCache.refresh([session]);
         return app._clientTeamFromCrud(crud, session);
       },
@@ -1996,11 +1993,11 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
         return useMemo(() => teams.map((t) => app._serverTeamFromCrud(t)), [teams]);
       },
       // END_PLATFORM react-like
-      createTeam: async (data: ServerTeamCreateOptions) => {
-        const team =  await app._interface.createServerTeam({
-          ...serverTeamCreateOptionsToCrud(data),
-          creator_user_id: crud.id,
-        });
+      createTeam: async (data: Omit<ServerTeamCreateOptions, "creatorUserId">) => {
+        const team = await app._interface.createServerTeam(serverTeamCreateOptionsToCrud({
+          creatorUserId: crud.id,
+          ...data,
+        }));
         await app._serverTeamsCache.refresh([undefined]);
         return app._serverTeamFromCrud(team);
       },
@@ -2977,6 +2974,8 @@ type ServerBaseUser = {
   setServerMetadata(metadata: any): Promise<void>,
   setClientReadOnlyMetadata(metadata: any): Promise<void>,
 
+  createTeam(data: Omit<ServerTeamCreateOptions, "creatorUserId">): Promise<ServerTeam>,
+
   // NEXT_LINE_PLATFORM react-like
   useContactChannels(): ServerContactChannel[],
   listContactChannels(): Promise<ServerContactChannel[]>,
@@ -3039,6 +3038,9 @@ type ServerUserCreateOptions = {
   otpAuthEnabled?: boolean,
   displayName?: string,
   primaryEmailVerified?: boolean,
+  clientMetadata?: any,
+  clientReadOnlyMetadata?: any,
+  serverMetadata?: any,
 }
 function serverUserCreateOptionsToCrud(options: ServerUserCreateOptions): UsersCrud["Server"]["Create"] {
   return {
@@ -3048,6 +3050,9 @@ function serverUserCreateOptionsToCrud(options: ServerUserCreateOptions): UsersC
     primary_email_auth_enabled: options.primaryEmailAuthEnabled,
     display_name: options.displayName,
     primary_email_verified: options.primaryEmailVerified,
+    client_metadata: options.clientMetadata,
+    client_read_only_metadata: options.clientReadOnlyMetadata,
+    server_metadata: options.serverMetadata,
   };
 }
 
@@ -3346,10 +3351,11 @@ export type TeamCreateOptions = {
   displayName: string,
   profileImageUrl?: string,
 }
-function teamCreateOptionsToCrud(options: TeamCreateOptions): TeamsCrud["Client"]["Create"] {
+function teamCreateOptionsToCrud(options: TeamCreateOptions, creatorUserId: string): TeamsCrud["Client"]["Create"] {
   return {
     display_name: options.displayName,
     profile_image_url: options.profileImageUrl,
+    creator_user_id: creatorUserId,
   };
 }
 
@@ -3382,9 +3388,15 @@ export type ServerListUsersOptions = {
   query?: string,
 };
 
-export type ServerTeamCreateOptions = TeamCreateOptions;
+export type ServerTeamCreateOptions = TeamCreateOptions & {
+  creatorUserId?: string,
+};
 function serverTeamCreateOptionsToCrud(options: ServerTeamCreateOptions): TeamsCrud["Server"]["Create"] {
-  return teamCreateOptionsToCrud(options);
+  return {
+    display_name: options.displayName,
+    profile_image_url: options.profileImageUrl,
+    creator_user_id: options.creatorUserId,
+  };
 }
 
 export type ServerTeamUpdateOptions = TeamUpdateOptions & {