+++ vitest.workspace.ts
@@ -1,6 +1,8 @@
-export default [
+import { defineWorkspace } from 'vitest/config';
+
+export default defineWorkspace([
   'packages/*',
   'apps/*',
   'examples/*',
   'docs',
-];
+]);