+++ packages/template/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack
 
+## 2.7.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.15
+  - @stackframe/stack-ui@2.7.15
+  - @stackframe/stack-sc@2.7.15
+
 ## 2.7.14
 
 ### Patch Changes